powerpc: make fadump resilient with memory add/remove events
Due to changes in memory resources caused by either memory hotplug or
memory online/offline events, the elfcorehdr, which describes the CPUs
and memory of the crashed kernel to the kernel that collects the dump,
becomes outdated. Consequently, attempting dump collection with an
outdated elfcorehdr can lead to failed or inaccurate dump collection.

The current solution to address the aforementioned issue is as follows:
Memory add/remove events are monitored in userspace using udev rules,
and whenever there are changes in memory resources, fadump is
re-registered, leading to the creation of a new elfcorehdr. This new
elfcorehdr includes the updated system memory information.

There are several notable issues associated with re-registering fadump
during memory add/remove events.

1. Performing bulk memory add/remove with udev-based fadump
   re-registration can lead to race conditions and, more importantly, it
   creates a wide window during which fadump is inactive until all
   memory events are settled.
2. Re-registering FADump for every memory add/remove event is
   inefficient.
3. Memory for elfcorehdr is allocated based on the memblock regions
   available during early boot and remains fixed thereafter. However, if
   the elfcorehdr is later recreated with additional memblock regions,
   its size will increase, potentially leading to memory corruption.

This patch addresses the aforementioned challenges by shifting the
creation of elfcorehdr from the first kernel (also referred to as the
crashed kernel), where it was created and frequently recreated for every
memory add/remove event, to the fadump kernel. As a result, the
elfcorehdr only needs to be created once, thus eliminating the necessity
to re-register fadump during memory add/remove events.

At present, the first kernel creates the elfcorehdr within the reserved
fadump area and includes its starting address in the elfcorehdr_addr
attribute of the fadump_crash_info_header structure. Now when the first
kernel crashes and fadump kernel boots, a platform-specific function is
called from setup_fadump/fadump.c to do the followings.

1. Sanity check for fadump_crash_info_header.
2. Update CPU notes in elfcorehdr.
3. Set the global variable elfcorehdr_addr to the address of the fadump
   crash info header's elfcorehdr. For vmcore module to process it
   later on.

The setup_fadump/fadump.c is updated to create elfcorehdr in fadump
kernel, along validating fadump_crash_info_header and setting
elfcorehdr_addr global variable which was previously done in platform
specific function. The updated setup_fadump/fadump.c do the followings:

1. Sanity check for fadump_crash_info_header.
2. create elfcorehdr and set its address to elfcorehdr_addr attribute of
   the fadump_crash_info_header structure.
3. A platform-specific function is called to update CPU notes in the
   elfcorehdr.
4. The global variable elfcorehdr_addr is set to the elfcorehdr's
   address. For vmcore module to process it later on.

The section below discusses the information required to create the
elfcorehdr and the changes made to make it available to the fadump
kernel if it's not already.

To create elfcorehdr, the following crashed kernel information is needed,
such as CPU notes, vmcoreinfo, and memory ranges.

At present, the CPU notes are already generated in the fadump kernel, so
no modification is needed in that regard. However, new members are added
to the `fadump_crash_info_header` structure to pass the crash kernel's
vmcoreinfo address and its size to the fadump kernel.

The fadump kernel has access to the memory ranges of the crashed kernel,
but the firmware relocates the crashed kernel's boot memory areas to a
predetermined fadump reserved area. This is done to ensure that the
fadump kernel can boot from its default memory location while retaining
the boot memory areas of the crashed kernel. To prepare the program
headers for the boot memory areas of the crashed kernel, the fadump
kernel needs to know their relocated addresses. To avail the relocated
boot memory address with the fadump kernel, a new member is added to
fw_dump structure. This is populated during early boot by the platform's
fadump init function.

The upcoming section describes what happens if either the crashed kernel
or the fadump kernel does not have the changes introduced here.

Let's consider the "old kernel" as the kernel without this patch and the
"new kernel" as the kernel with this patch included.

Scenario 1:
-----------
When the old kernel crashes and a new kernel is booted as the fadump
kernel, the dump collection succeeds. This is achieved by updating the
fadump crash info header with a new magic number that helps identify
the crashed kernel as either the old or new kernel. In this scenario,
as the crashed kernel is the old kernel, the elfcorehdr must have been
prepared beforehand; hence, the elfcorehdr creation is skipped in the
fadump kernel.

Scenario 2:
----------
When the new kernel crashed and old kernel is booted as the fadump
kernel, the dump collection will fail. Because old kernel cannot process
fadump crash info header with new magic number.

To maintain backward compatibility and prevent future changes to magic
number of fadump crash info header, a version field has been introduced
to the structure.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>