kexec/crash: no crash update when kexec in progress
The following errors are observed when kexec is done with SMT=off on
powerpc.

[  358.458385] Removing IBM Power 842 compression device
[  374.795734] kexec_core: Starting new kernel
[  374.795748] kexec: Waking offline cpu 1.
[  374.875695] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  374.935833] kexec: Waking offline cpu 2.
[  375.015664] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  375.075820] kexec: Waking offline cpu 3.
[  375.155665] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  375.215790] kexec: Waking offline cpu 4.
[  375.305665] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  375.365808] kexec: Waking offline cpu 5.
[  375.455665] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  375.515823] kexec: Waking offline cpu 6.
[  375.635667] crash hp: kexec_trylock() failed, elfcorehdr may be inaccurate
[  375.695836] kexec: Waking offline cpu 7.

During kexec, the offline CPUs are brought online, which triggers the
crash hotplug handler `crash_handle_hotplug_event()` to update the kdump
image. Given that the system is on the kexec path and the kexec lock is
taken, the `crash_handle_hotplug_event()` function fails to take the
same lock to update the kdump image, resulting in the above error
messages.

To fix this, let's return from `crash_handle_hotplug_event()` if kexec
is in progress.

The same applies to the `crash_check_hotplug_support()` function.
Return 0 if kexec is in progress.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>
Cc: Baoquan He <bhe@redhat.com>
Cc: Hari Bathini <hbathini@linux.ibm.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: kexec@lists.infradead.org
Cc: linuxppc-dev@ozlabs.org
Cc: linux-kernel@vger.kernel.org
Cc: x86@kernel.org